requirements.txt
transformers
torch 
torchvision 
torchaudio
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
datetime